Function that create Excursion is finished Now it is time to create an offer

+++ AGP/src/business/ActivityPlace.java
@@ -2,7 +2,7 @@ package business;
 
 public class ActivityPlace extends Place{
 
-	public ActivityPlace(int id, String name, float visitDuration, float entrancePrice, String description, Coordinates coordinates, Island island) {
+	public ActivityPlace(int id, String name, int visitDuration, float entrancePrice, String description, Coordinates coordinates, Island island) {
 		super(id, name, visitDuration, entrancePrice, description, coordinates, island);
 	}
 }
